+++ packages/camera/camera/example/lib/main.dart
@@ -49,6 +49,8 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
   Animation<double> _flashModeControlRowAnimation;
   AnimationController _exposureModeControlRowAnimationController;
   Animation<double> _exposureModeControlRowAnimation;
+  AnimationController _focusModeControlRowAnimationController;
+  Animation<double> _focusModeControlRowAnimation;
   double _minAvailableZoom;
   double _maxAvailableZoom;
   double _currentScale = 1.0;
@@ -75,6 +77,14 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
     );
     _exposureModeControlRowAnimation = CurvedAnimation(
       parent: _exposureModeControlRowAnimationController,
+      curve: Curves.easeInCubic,
+    );
+    _focusModeControlRowAnimationController = AnimationController(
+      duration: const Duration(milliseconds: 300),
+      vsync: this,
+    );
+    _focusModeControlRowAnimation = CurvedAnimation(
+      parent: _focusModeControlRowAnimationController,
       curve: Curves.easeInCubic,
     );
   }
@@ -161,18 +171,18 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
         ),
       );
     } else {
-      return AspectRatio(
-        aspectRatio: controller.value.aspectRatio,
-        child: Listener(
-          onPointerDown: (_) => _pointers++,
-          onPointerUp: (_) => _pointers--,
+      return Listener(
+        onPointerDown: (_) => _pointers++,
+        onPointerUp: (_) => _pointers--,
+        child: CameraPreview(
+          controller,
           child: LayoutBuilder(
               builder: (BuildContext context, BoxConstraints constraints) {
             return GestureDetector(
+              behavior: HitTestBehavior.opaque,
               onScaleStart: _handleScaleStart,
               onScaleUpdate: _handleScaleUpdate,
               onTapDown: (details) => onViewFinderTap(details, constraints),
-              child: CameraPreview(controller),
             );
           }),
         ),
@@ -250,14 +260,29 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
                   controller != null ? onExposureModeButtonPressed : null,
             ),
             IconButton(
+              icon: Icon(Icons.filter_center_focus),
+              color: Colors.blue,
+              onPressed: controller != null ? onFocusModeButtonPressed : null,
+            ),
+            IconButton(
               icon: Icon(enableAudio ? Icons.volume_up : Icons.volume_mute),
               color: Colors.blue,
               onPressed: controller != null ? onAudioModeButtonPressed : null,
+            ),
+            IconButton(
+              icon: Icon(controller?.value?.isCaptureOrientationLocked ?? false
+                  ? Icons.screen_lock_rotation
+                  : Icons.screen_rotation),
+              color: Colors.blue,
+              onPressed: controller != null
+                  ? onCaptureOrientationLockButtonPressed
+                  : null,
             ),
           ],
         ),
         _flashModeControlRowWidget(),
         _exposureModeControlRowWidget(),
+        _focusModeControlRowWidget(),
       ],
     );
   }
@@ -313,6 +338,17 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
   }
 
   Widget _exposureModeControlRowWidget() {
+    final ButtonStyle styleAuto = TextButton.styleFrom(
+      primary: controller?.value?.exposureMode == ExposureMode.auto
+          ? Colors.orange
+          : Colors.blue,
+    );
+    final ButtonStyle styleLocked = TextButton.styleFrom(
+      primary: controller?.value?.exposureMode == ExposureMode.locked
+          ? Colors.orange
+          : Colors.blue,
+    );
+
     return SizeTransition(
       sizeFactor: _exposureModeControlRowAnimation,
       child: ClipRect(
@@ -327,12 +363,9 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
                 mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                 mainAxisSize: MainAxisSize.max,
                 children: [
-                  FlatButton(
+                  TextButton(
                     child: Text('AUTO'),
-                    textColor:
-                        controller?.value?.exposureMode == ExposureMode.auto
-                            ? Colors.orange
-                            : Colors.blue,
+                    style: styleAuto,
                     onPressed: controller != null
                         ? () =>
                             onSetExposureModeButtonPressed(ExposureMode.auto)
@@ -342,12 +375,9 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
                       showInSnackBar('Resetting exposure point');
                     },
                   ),
-                  FlatButton(
+                  TextButton(
                     child: Text('LOCKED'),
-                    textColor:
-                        controller?.value?.exposureMode == ExposureMode.locked
-                            ? Colors.orange
-                            : Colors.blue,
+                    style: styleLocked,
                     onPressed: controller != null
                         ? () =>
                             onSetExposureModeButtonPressed(ExposureMode.locked)
@@ -374,6 +404,59 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
                         : setExposureOffset,
                   ),
                   Text(_maxAvailableExposureOffset.toString()),
+                ],
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _focusModeControlRowWidget() {
+    final ButtonStyle styleAuto = TextButton.styleFrom(
+      primary: controller?.value?.focusMode == FocusMode.auto
+          ? Colors.orange
+          : Colors.blue,
+    );
+    final ButtonStyle styleLocked = TextButton.styleFrom(
+      primary: controller?.value?.focusMode == FocusMode.locked
+          ? Colors.orange
+          : Colors.blue,
+    );
+
+    return SizeTransition(
+      sizeFactor: _focusModeControlRowAnimation,
+      child: ClipRect(
+        child: Container(
+          color: Colors.grey.shade50,
+          child: Column(
+            children: [
+              Center(
+                child: Text("Focus Mode"),
+              ),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                mainAxisSize: MainAxisSize.max,
+                children: [
+                  TextButton(
+                    child: Text('AUTO'),
+                    style: styleAuto,
+                    onPressed: controller != null
+                        ? () => onSetFocusModeButtonPressed(FocusMode.auto)
+                        : null,
+                    onLongPress: () {
+                      if (controller != null) controller.setFocusPoint(null);
+                      showInSnackBar('Resetting focus point');
+                    },
+                  ),
+                  TextButton(
+                    child: Text('LOCKED'),
+                    style: styleLocked,
+                    onPressed: controller != null
+                        ? () => onSetFocusModeButtonPressed(FocusMode.locked)
+                        : null,
+                  ),
                 ],
               ),
             ],
@@ -468,10 +551,12 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
   }
 
   void onViewFinderTap(TapDownDetails details, BoxConstraints constraints) {
-    controller.setExposurePoint(Offset(
+    final offset = Offset(
       details.localPosition.dx / constraints.maxWidth,
       details.localPosition.dy / constraints.maxHeight,
-    ));
+    );
+    controller.setExposurePoint(offset);
+    controller.setFocusPoint(offset);
   }
 
   void onNewCameraSelected(CameraDescription cameraDescription) async {
@@ -482,6 +567,7 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
       cameraDescription,
       ResolutionPreset.medium,
       enableAudio: enableAudio,
+      imageFormatGroup: ImageFormatGroup.jpeg,
     );
 
     // If the controller is updated then update the UI.
@@ -494,10 +580,16 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
 
     try {
       await controller.initialize();
-      _minAvailableExposureOffset = await controller.getMinExposureOffset();
-      _maxAvailableExposureOffset = await controller.getMaxExposureOffset();
-      _maxAvailableZoom = await controller.getMaxZoomLevel();
-      _minAvailableZoom = await controller.getMinZoomLevel();
+      await Future.wait([
+        controller
+            .getMinExposureOffset()
+            .then((value) => _minAvailableExposureOffset = value),
+        controller
+            .getMaxExposureOffset()
+            .then((value) => _maxAvailableExposureOffset = value),
+        controller.getMaxZoomLevel().then((value) => _maxAvailableZoom = value),
+        controller.getMinZoomLevel().then((value) => _minAvailableZoom = value),
+      ]);
     } on CameraException catch (e) {
       _showCameraException(e);
     }
@@ -526,6 +618,7 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
     } else {
       _flashModeControlRowAnimationController.forward();
       _exposureModeControlRowAnimationController.reverse();
+      _focusModeControlRowAnimationController.reverse();
     }
   }
 
@@ -535,6 +628,17 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
     } else {
       _exposureModeControlRowAnimationController.forward();
       _flashModeControlRowAnimationController.reverse();
+      _focusModeControlRowAnimationController.reverse();
+    }
+  }
+
+  void onFocusModeButtonPressed() {
+    if (_focusModeControlRowAnimationController.value == 1) {
+      _focusModeControlRowAnimationController.reverse();
+    } else {
+      _focusModeControlRowAnimationController.forward();
+      _flashModeControlRowAnimationController.reverse();
+      _exposureModeControlRowAnimationController.reverse();
     }
   }
 
@@ -542,6 +646,19 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
     enableAudio = !enableAudio;
     if (controller != null) {
       onNewCameraSelected(controller.description);
+    }
+  }
+
+  void onCaptureOrientationLockButtonPressed() async {
+    if (controller != null) {
+      if (controller.value.isCaptureOrientationLocked) {
+        await controller.unlockCaptureOrientation();
+        showInSnackBar('Capture orientation unlocked');
+      } else {
+        await controller.lockCaptureOrientation();
+        showInSnackBar(
+            'Capture orientation locked to ${controller.value.lockedCaptureOrientation.toString().split('.').last}');
+      }
     }
   }
 
@@ -556,6 +673,13 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
     setExposureMode(mode).then((_) {
       if (mounted) setState(() {});
       showInSnackBar('Exposure mode set to ${mode.toString().split('.').last}');
+    });
+  }
+
+  void onSetFocusModeButtonPressed(FocusMode mode) {
+    setFocusMode(mode).then((_) {
+      if (mounted) setState(() {});
+      showInSnackBar('Focus mode set to ${mode.toString().split('.').last}');
     });
   }
 
@@ -672,6 +796,15 @@ class _CameraExampleHomeState extends State<CameraExampleHome>
     });
     try {
       offset = await controller.setExposureOffset(offset);
+    } on CameraException catch (e) {
+      _showCameraException(e);
+      rethrow;
+    }
+  }
+
+  Future<void> setFocusMode(FocusMode mode) async {
+    try {
+      await controller.setFocusMode(mode);
     } on CameraException catch (e) {
       _showCameraException(e);
       rethrow;