+++ packages/camera/camera_platform_interface/lib/src/platform_interface/camera_platform.dart
@@ -6,10 +6,13 @@ import 'dart:async';
 import 'dart:math';
 
 import 'package:camera_platform_interface/camera_platform_interface.dart';
+import 'package:camera_platform_interface/src/events/device_event.dart';
 import 'package:camera_platform_interface/src/method_channel/method_channel_camera.dart';
 import 'package:camera_platform_interface/src/types/exposure_mode.dart';
+import 'package:camera_platform_interface/src/types/focus_mode.dart';
 import 'package:camera_platform_interface/src/types/image_format_group.dart';
 import 'package:cross_file/cross_file.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
@@ -84,6 +87,27 @@ abstract class CameraPlatform extends PlatformInterface {
     throw UnimplementedError('onCameraError() is not implemented.');
   }
 
+  /// The device orientation changed.
+  ///
+  /// Implementations for this:
+  /// - Should support all 4 orientations.
+  /// - Should not emit new values when the screen orientation is locked.
+  Stream<DeviceOrientationChangedEvent> onDeviceOrientationChanged() {
+    throw UnimplementedError(
+        'onDeviceOrientationChanged() is not implemented.');
+  }
+
+  /// Locks the capture orientation.
+  Future<void> lockCaptureOrientation(
+      int cameraId, DeviceOrientation orientation) {
+    throw UnimplementedError('lockCaptureOrientation() is not implemented.');
+  }
+
+  /// Unlocks the capture orientation.
+  Future<void> unlockCaptureOrientation(int cameraId) {
+    throw UnimplementedError('unlockCaptureOrientation() is not implemented.');
+  }
+
   /// Captures an image and returns the file where it was saved.
   Future<XFile> takePicture(int cameraId) {
     throw UnimplementedError('takePicture() is not implemented.');
@@ -129,7 +153,7 @@ abstract class CameraPlatform extends PlatformInterface {
     throw UnimplementedError('setExposureMode() is not implemented.');
   }
 
-  /// Sets the exposure point for automatically determining the exposure value.
+  /// Sets the exposure point for automatically determining the exposure values.
   Future<void> setExposurePoint(int cameraId, Point<double> point) {
     throw UnimplementedError('setExposurePoint() is not implemented.');
   }
@@ -164,6 +188,16 @@ abstract class CameraPlatform extends PlatformInterface {
   /// Returns the (rounded) offset value that was set.
   Future<double> setExposureOffset(int cameraId, double offset) {
     throw UnimplementedError('setExposureOffset() is not implemented.');
+  }
+
+  /// Sets the focus mode for taking pictures.
+  Future<void> setFocusMode(int cameraId, FocusMode mode) {
+    throw UnimplementedError('setFocusMode() is not implemented.');
+  }
+
+  /// Sets the focus point for automatically determining the focus values.
+  Future<void> setFocusPoint(int cameraId, Point<double> point) {
+    throw UnimplementedError('setFocusPoint() is not implemented.');
   }
 
   /// Gets the maximum supported zoom level for the selected camera.