+++ packages/google_sign_in/google_sign_in_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.0-nullsafety
+
+* Migration to nnbd.
+
 ## 1.1.3
 
 * Update Flutter SDK constraint.